Added getters for the different parameters
